fix(exporter/prometheus): move synctest to stable API, drop build tag
Merge background cleanup test into prometheus_test.go using
synctest.Test (stable in Go 1.25+) instead of the experimental
synctest.Run. Remove the separate test file and goexperiment build tag.
Strip unnecessary comments.
